+++ src/utils/times.js
@@ -5,4 +5,9 @@ const minute = second * 60
 const hour = minute * 60
 const day = hour * 24
 
-module.exports = day
+module.exports = {
+	second,
+	minute,
+	hour,
+	day
+}
