--- test/utils/day.js
@@ -1,11 +1,0 @@
-'use strict'
-
-const test = require('ava')
-
-const day = require('../../src/utils/day')
-
-test('return one day in milliseconds', async (t) => {
-
-	t.is(day, 86400000)
-
-})
